requirements.txt
Django
psycopg2-binary
djangorestframework
django-environ
django-unixtimestampfield
drf-yasg
django-cors-headers
djangorestframework-jsonapi
django-filter